requirements.txt
git+ssh://github.com/poddile/mscli#egg=mscli
pywebview